rendering 'material' working, player movement and camera working, hpx sim thread working

+++ game/Syris/src/Syris/renderer/IndexBuffer.hpp
@@ -6,7 +6,7 @@ namespace Syris{
     class IndexBuffer{
         public:
             struct CreateInfo{
-                uint32_t indices_count;
+                uint64_t indices_count;
                 uint32_t* indices;
                 bool dynamic;
             };
